scc modals connected to new proposal flow
